Fixes for  leaving the spline on 180 interruptions
if move target was 180 from the current momentum the effective error would be 0 and the character would just continue in the wrong direction. Added a minimum error so that a small amount of the curve tangent will always get incorporated into the trajectory. Also reordered the evaluation cycle so that after updating segments we always try at least once to get a valid rabbit position